+++ forms-flow-web/src/components/Draft/DraftFilter.js
@@ -84,8 +84,7 @@ const DraftFilter = ({ setDisplayFilter, filterParams, setFilterParams }) => {
 
   return (
     <div
-      className="Filter-listview "
-      style={{ minWidth: "500px" }}
+      className="appFilter-list-view"
       ref={createSearchNode}
     >