+++ forms-flow-web/src/components/Form/constants/ClientTable.js
@@ -106,9 +106,7 @@ function ClientTable() {
         <tr>
           <td colSpan="3">
             <div
-              className="d-flex align-items-center justify-content-center flex-column w-100"
-              style={{ minHeight: "300px" }}
-            >
+              className="d-flex align-items-center justify-content-center clientForm-table-col flex-column w-100">
               <h3>{t("No forms found")}</h3>
               <p>{t("Please change the selected filters to view Forms")}</p>
             </div>
@@ -139,7 +137,7 @@ function ClientTable() {
   return (
     <>
       <LoadingOverlay active={searchFormLoading} spinner text={t("Loading...")}>
-        <div style={{ minHeight: "400px" }}>
+        <div className="min-height-400">
           <table className="table custom-table table-responsive-sm">
             <thead>
               <tr>
@@ -148,33 +146,25 @@ function ClientTable() {
                     <span>{t("Form Title")}</span>
                     <span>
                       {isAscending ? (
-                        <i
+                        <i 
                           data-testid="form-desc-sort-icon"
-                          className="fa fa-sort-alpha-asc ms-2"
+                          className="fa fa-sort-alpha-asc ms-2 cursor-pointer fs-16"
                           onClick={() => {
                             updateSort("desc");
                           }}
                           data-toggle="tooltip"
-                          title={t("Descending")}
-                          style={{
-                            cursor: "pointer",
-                            fontSize: "16px",
-                          }}
-                        ></i>
+                          title={t("Descending")}>
+                        </i>
                       ) : (
                         <i
                           data-testid="form-asc-sort-icon"
-                          className="fa fa-sort-alpha-desc ms-2"
+                          className="fa fa-sort-alpha-desc ms-2 cursor-pointer fs-16"
                           onClick={() => {
                             updateSort("asc");
                           }}
                           data-toggle="tooltip"
-                          title={t("Ascending")}
-                          style={{
-                            cursor: "pointer",
-                            fontSize: "16px",
-                          }}
-                        ></i>
+                          title={t("Ascending")}>
+                        </i>
                       )}
                     </span>
                   </div>
@@ -190,9 +180,9 @@ function ClientTable() {
                       onKeyDown={(e) =>
                         e.keyCode === 13 ? handleSearch() : ""
                       }
+                      className="bg-white"
                       data-testid="form-search-input-box"
                       placeholder={t("Search by form title")}
-                      style={{ backgroundColor: "#ffff" }}
                       title={t("Search by form title")}
                     />
                     {search && (
@@ -206,12 +196,12 @@ function ClientTable() {
                       </InputGroup.Append>
                     )}
                     <InputGroup.Append
+                      className="cursor-pointer"
                       onClick={handleSearch}
                       data-testid="form-search-click-button"
                       disabled={!search?.trim()}
-                      style={{ cursor: "pointer" }}
                     >
-                      <InputGroup.Text style={{ backgroundColor: "#ffff" }} className="h-100">
+                      <InputGroup.Text className="h-100 bg-white">
                         <i className="fa fa-search"></i>
                       </InputGroup.Text>
                     </InputGroup.Append>
@@ -251,11 +241,7 @@ function ClientTable() {
                       </td>
                       <td
                         data-testid={`form-description${e._id}`}
-                        className="text-truncate"
-                        style={{
-                          maxWidth: "350px",
-                        }}
-                      >
+                        className="text-truncate">
                         {extractContent(e.description)}
                       </td>
 
@@ -279,7 +265,6 @@ function ClientTable() {
                             </h4>
 
                             <div
-                              style={{ maxWidth: "68vw" }}
                               className="form-description-p-tag "
                               dangerouslySetInnerHTML={{
                                 __html: sanitize(e?.description, {