+++ forms-flow-web/src/components/Insights/Insights.js
@@ -140,23 +140,11 @@ const Insights = React.memo((props) => {
             {activeDashboard.public_url ? (
               <iframe
                 title="dashboard"
-                style={{
-                  width: "100%",
-                  height: "auto",
-                  overflow: "visible",
-                  border: "none",
-                  minHeight: "60vh",
-                }}
+                className="w-100 h-auto min-vh-60 overflow-visible border-none"
                 src={activeDashboard.public_url}
               />
             ) : !isDashboardDetailUpdated ? (
-              <div
-                style={{
-                  position:'absolute',
-                  left: '52%',
-                  marginTop: '400px',
-                  transform: 'translate(-50%, -90%)',
-                }}>
+              <div className="position-absolute no-dashboard-detail">
                 <Loading />
               </div>
             ) : (